+++ tools/apidiff/cmd/changelog.go
@@ -9,8 +9,8 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/Azure/azure-sdk-for-go/tools/apidiff/report"
 	"github.com/Azure/azure-sdk-for-go/tools/common/markdown"
+	"github.com/Azure/azure-sdk-for-go/tools/common/report"
 	"github.com/spf13/cobra"
 )
