+++ tools/apidiff/cmd/packages.go
@@ -10,9 +10,9 @@ import (
 	"os"
 	"strings"
 
-	"github.com/Azure/azure-sdk-for-go/tools/apidiff/report"
 	"github.com/Azure/azure-sdk-for-go/tools/common/exports"
 	"github.com/Azure/azure-sdk-for-go/tools/common/repo"
+	"github.com/Azure/azure-sdk-for-go/tools/common/report"
 	"github.com/spf13/cobra"
 )
